==> src/zarni/02_InitialModels.Rmd <==
---
title: "02_InitialModels"
author: "Zarni Htet"
date: "March 7, 2018"
output: github_document
---
### Introduction

The goal of this code file is to explore whether infant media exposure is associated with weight/bmi trajectories during their infant to early childhood periods. We explore a simple linear regression model and a multilevel regression model in the code file.

### Datasets Involved

The linearly interpolated final data set generated from *01a_Linear_Interpolation* is used here.

### Admnistration

Professor Marc Scott and Professor Daphna Harel are the supervisors of this project. The data is from the Belle Lab at the Bellevue Hospital. Additional background on the project is in the *README* at the root directory of the Github repository associated with the project.

#### R Libraries

This block has all the *required* libraries for this code file.

```{r, R.options=FALSE, warning=FALSE, message=FALSE}
library(lme4) # For lmer function
library(lmerTest)
library(rio) # For importing data frames
```

#### Importing the data frames to be used
```{r}
bmi_media <- import("../../data/final/final_interp_data.csv")
```

### Functional Form

#### Simple Linear Regression

We are regressing BMI on Media and Time.

$$
Y(BMI) \sim X (Media)+ t (Months) + \epsilon   
$$

```{r}
fit <- lm(zBMI ~ Media + Months, data = bmi_media)
print(summary(fit))
```

#### Simple Linear Regression Interpretation

On average, a unit increase in media is associated with a 0.02868 increase in BMI. 

#### Multi Level Regression 

##### Varying Intercept Model with individual predictor

$$
BMI_{ti} = b_{0} + b_{1}Media_{ti} + b_{2}Months_{ti} + \zeta_{i} + \epsilon_{ti}
$$
```{r}
#lmer code
M1 = lmer(zBMI~Media + Months + (1|ID), data = bmi_media)
print(summary(M1))
```

#### Model Interpretation

Controlling for differences between subjects, the effect of one unit change in Media is 0.0161 to BMI and it is not statistically significant assuming our model assumptions are correct. With the likelihood ratio test below, the random effects are warranted.

#### Likelihood ratio test

```{r}
print(rand(M1))
```
